divided the component into folders, added drag-and-drop
