+++ src/components/ui/admin/settings/listSlides/hooks/useCreateImageOptions.tsx
@@ -1,0 +1,13 @@
+import { useGetImages } from '~hooks/images';
+
+export const useCreateImageOptions = () => {
+    const { data: images } = useGetImages();
+
+    return images?.map((image) => {
+        return {
+            label: <img src={image.url} alt="" />,
+            value: image.url,
+            base64: image.base64,
+        };
+    });
+};
