+++ src/components/ui/admin/settings/listSlides/index.tsx
@@ -1,0 +1,1 @@
+export * from './list';
