+++ src/components/ui/admin/settings/listSlides/types/index.ts
@@ -1,0 +1,8 @@
+export type Slide = {
+    id: number;
+    img?: string;
+    title?: string;
+    desc?: string;
+    active: boolean;
+    base64: string;
+};
